Wrappers makes it a wrap
Centralized helpers in UOXJSClasses.h.
Removed duplicate implementations from SEFunctions.cpp and UOXJSMethods.cpp.
GetBaseObject() checks items first, then characters.
Uses toObjectOrNull().
Retrieves the shared pointer as CBaseObject *.
